requirements.txt
# Enterprise Bot - Lean Stack
# No ML pipeline bloat, no Supabase wrapper

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP & WebSocket
httpx>=0.27.0
websockets>=12.0
aiohttp>=3.8.0

# Database (Raw Postgres + pgvector)
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5

# Document Processing
python-docx>=1.1.0
python-multipart>=0.0.6
requests>=2.31.0
rank-bm25>=0.2.2

# Config
python-dotenv>=1.0.0
PyYAML>=6.0

# Grok API (uses OpenAI-compatible client)
openai>=1.12.0
xai-sdk>=1.4.0

redis>=5.0.0          # Async Redis client (has aioredis built-in now)
hiredis>=2.0.0        # Fast Redis parser (optional but recommended)

# Azure AD / MSAL
msal>=1.24.0
PyJWT>=2.8.0
cryptography>=41.0.0

# SDK Tools (for Railway, Memory, DB tools)
# Note: python-dotenv already listed above
faiss-cpu>=1.7.4  # or faiss-gpu if CUDA available
numpy>=1.24.0  # required by FAISS

# Observability (Phase 1)
psutil>=5.9.0

# Credit Module (PDF generation, SQL Server)
reportlab>=4.0.0
pyodbc>=5.0.0